+++ internal/utils/helpers/helpers.go
@@ -80,7 +80,7 @@ func FormatUint64(s string, defaultValue uint64) uint64 {
 }
 
 // FormatAmount is used to convert a uint256 string to a float64, based on the provided decimal
-func FormatAmount(balanceToken string, decimals int) (float64, *bigNumber.BigFloat) {
+func FormatAmount(balanceToken string, decimals int) (float64, *bigNumber.Float) {
 	fTotalAssets := bigNumber.NewFloat().SetString(balanceToken)
 	fDecimals := bigNumber.NewFloat(math.Pow10(decimals))
 	humanizedBalance := bigNumber.NewFloat().Quo(fTotalAssets, fDecimals)
